+++ src/tests/trivial_test.cc
@@ -78,31 +78,9 @@ static void parentTest() {
     wchar_t cmdline[1024];
     GetModuleFileNameW(nullptr, program, 1024);
 
-    {
-        // XXX: We'd like to use swprintf, which is part of C99 and takes a
-        // size_t maxlen argument.  MinGW-w64 has this function, as does MSVC.
-        // The old MinGW doesn't, though -- instead, it apparently provides an
-        // swprintf taking no maxlen argument.  This *might* be a regression?
-        // (There is also no swnprintf, but that function is obsolescent with a
-        // correct swprintf, and it isn't in POSIX or ISO C.)
-        //
-        // Visual C++ 6 also provided this non-conformant swprintf, and I'm
-        // guessing MSVCRT.DLL does too.  (My impression is that the old MinGW
-        // prefers to rely on MSVCRT.DLL for convenience?)
-        //
-        // I could compile differently for old MinGW, but what if it fixes its
-        // function later?  Instead, use a workaround.  It's starting to make
-        // sense to drop MinGW support in favor of MinGW-w64.  This is too
-        // annoying.
-        //
-        // grepbait: OLD-MINGW / WINPTY_TARGET_MSYS1
-        cmdline[0] = L'\0';
-        wcscat(cmdline, L"\"");
-        wcscat(cmdline, program);
-        wcscat(cmdline, L"\" CHILD");
-    }
-    // swnprintf(cmdline, sizeof(cmdline) / sizeof(cmdline[0]),
-    //           L"\"%ls\" CHILD", program);
+    const int commandLength = swprintf_s(
+        cmdline, _countof(cmdline), L"\"%ls\" CHILD", program);
+    assert(commandLength > 0);
 
     auto agentCfg = winpty_config_new(0, nullptr);
     assert(agentCfg != nullptr);